+++ docs/changelog.rst
@@ -15,6 +15,7 @@ Internal Changes
   Brindley`_.
 - Use the conda-forge recommended 'python-build' package instead of 'build', which is apparently deprecated as too
   general a name (:issue:`481`, :merge:`589`). By `Kyle Brindley`_.
+- Overwrite h5 files if content changed on parameter generators (:issue:`441`, :merge:`590`). By `Sergio Cordova`_
 
 *******************
 0.6.19 (2023-06-14)