+++ docs/changelog.rst
@@ -11,6 +11,8 @@ Changelog
 Documentation
 =============
 - Removed semaphore files in tutorials (:issue:`488`, :merge:`591`). By `Sergio Cordova`_
+- Updated parameter study CLI messages to reflect yaml file behavior changes (:issue:`490`, :merge:`593`). By `Sergio
+  Cordova`_
 
 Enhancement
 ===========