+++ docs/changelog.rst
@@ -26,6 +26,8 @@ Internal Changes
   Brindley`_.
 - Use the conda-forge recommended 'python-build' package instead of 'build', which is apparently deprecated as too
   general a name (:issue:`481`, :merge:`589`). By `Kyle Brindley`_.
+- Updated h5 and yaml parameter generator tests to use the same data input (:issue:`491`, :merge:`594`). By `Sergio
+  Cordova`_
 
 *******************
 0.6.19 (2023-06-14)